requirements.txt
openai>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
typing>=3.7.4.3